[EasyWebsite] FIX: BUG when accessing /admin and no token => no redirection.

+++ EasyWebsite.API/Scripts/app/services/AuthService.js
@@ -68,25 +68,25 @@
 
             var authData = localStorageService.get('authorizationData');
 
-            if (authData) {
+            if (authData && authData.useRefreshTokens) {
 
-                if (authData.useRefreshTokens) {
+                var data = "grant_type=refresh_token&refresh_token=" + authData.refreshToken + "&client_id=ngAuthApp";
 
-                    var data = "grant_type=refresh_token&refresh_token=" + authData.refreshToken + "&client_id=" + ngAuthSettings.clientId;
+                localStorageService.remove('authorizationData');
 
-                    localStorageService.remove('authorizationData');
+                $http = $http || $injector.get('$http');
+                $http.post('/token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
 
-                    $http.post('/token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
+                    localStorageService.set('authorizationData', { token: response.access_token, userName: response.userName, refreshToken: response.refresh_token, useRefreshTokens: true });
 
-                        localStorageService.set('authorizationData', { token: response.access_token, userName: response.userName, refreshToken: response.refresh_token, useRefreshTokens: true });
+                    deferred.resolve(response);
 
-                        deferred.resolve(response);
-
-                    }).error(function (err, status) {
-                        _logOut();
-                        deferred.reject(err);
-                    });
-                }
+                }).error(function (err, status) {
+                    _logOut();
+                    deferred.reject(err);
+                });
+            } else {
+                deferred.reject();
             }
 
             return deferred.promise;
